requirements.txt
typer
rich
questionary
python-dateutil
Pillow
